# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Restaurant API",
        "contact": {
            "name": "Hojune Yoo",
            "email": "soy0803@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/options": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get all menu options",
                "operationId": "GetAllOptions",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Save an option",
                "operationId": "SaveOption",
                "parameters": [
                    {
                        "name": "menu_id",
                        "in": "formData",
                        "description": "which menu",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Option name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Option description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "price of option",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/options/{optionId}": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get an option",
                "operationId": "GetOption",
                "parameters": [
                    {
                        "name": "optionId",
                        "in": "path",
                        "description": "Option id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Options"
                ],
                "summary": "Update option",
                "operationId": "updateOption",
                "parameters": [
                    {
                        "name": "optionId",
                        "in": "path",
                        "description": "Which Option",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Option name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "price of option",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Options"
                ],
                "summary": "Delete an option",
                "operationId": "DeleteOption",
                "parameters": [
                    {
                        "name": "optionId",
                        "in": "path",
                        "description": "Option id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get all messages",
                "operationId": "GetAllMessages",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Save a message",
                "operationId": "SaveMessage",
                "parameters": [
                    {
                        "name": "text",
                        "in": "formData",
                        "description": "message context",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "visit_id",
                        "in": "formData",
                        "description": "visit Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/messages/{messageId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get a Message",
                "operationId": "GetMessage",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "Message id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Delete a message",
                "operationId": "DeleteMessage",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "Message id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}